sample_container/requirements.txt
--extra-index-url https://pypi.ngc.nvidia.com
nvidia-pytriton==0.5.0
numpy
requests
pillow
git+https://github.com/NVIDIA/nv-cloud-function-helpers.git